moved themes to a separate directory; updated README.md
