+++ README.md
@@ -88,19 +88,66 @@ of completed tasks.
 - `eisenhower` contains the code for the UI of the app, specifically - the Eisenhower Matrix. _Undoubtedly the biggest
 mess here_.
 - `completed_tasks_view` contains the code for the UI of the list of completed tasks.
-- `misc` contains some other minor tools like the `StyleSwitcher` that allows users to switch between dark and light
-themes.
+- `themes` contains the code that controls the UI themes.
+- `misc` is for stuff that doesn't really fit anywhere else but doesn't deserve its own folder either, like the code for
+the button that opens the github page for the project.
 
 
 ## Style guide
 
-While adding new themes at the moment is a bit of a mess _(you have to hard-code a theme symbol element for the
-switch-theme button)_, you can modify the colors in an old one.
+A theme is simply a set of CSS rules that are applied to the page once the theme is selected. Adding a custom theme is
+simple:
+
+1. Select an existing theme to be used as a starting point. For example, use `web/dark_mode.css` if you're creating a
+dark mode theme; use `web/light_mode.css` if you're creating a light mode theme. Copy the file and place it in the
+`web/themes` directory alongside the other custom themes. Give it a nice name (the name should be a valid HTML class -
+no spaces, ideally use only letters and underscores to separate words).
+
+2. Add your theme to the `source/themes.ts` file. To do that, add an entry inside the `initCustomThemes` function
+(take a look at how the other themes are added there). You will have to add the following lines to this function:
+
+```ts
+themes.addTheme (new CustomTheme ({
+	themeName: 'Human readable name',
+	themeClass: 'the_name_of_your_theme_file_without_the_extension', 
+	themeSVGIcon: '<path d="TH3-SVG-1C0N T0-R3PR3S-3nt Y08R-7H3Me"/>'
+}));
+```
+
+Set `themeName` to the name of your theme, `themeClass` to the name of your css file (no extension, no full path). Use `themeSVGIcon`
+to select an SVG icon for your theme.
+
+You will have to rebuild the app for the theme to be selectable.
+
+3. Open the CSS file you created and make it so that the name of the class inside matched the name of the theme. For
+example, for a theme named `glitter_factory_explosion` the beginning of the `glitter_factory_explosion.css` file would
+look like this:
+
+```css
+
+.glitter_factory_explosion {
+	--accent-color: yellow;
+	--background-color: red;
+	--text-color: pink;
+	...
+```
+
+4. Modify the `.css` theme in a way you want!
+
+In order to load custom themes locally you'll have to run an HTTP server that serves files from the `web/` directory.
+While I am looking into how one can setup a web server using Node, you can start one using Python:
+
+```bash
+cd ./web
+python -m http.server
+```
+
+You will then be able to open the app at `localhost:8000` with the full support of custom themes.
 
 ### Theme structure
 
-A theme is just a CSS file (`web/light_mode.css`, `web/dark_mode.css`) with a set of `var(...)` variables defined in it,
-for example:
+A basic theme is just a CSS file (`web/light_mode.css`, `web/dark_mode.css`) with a set of `var(...)` variables defined
+in it, for example:
 
 ```css
 --accent-color: #10a810;
@@ -110,3 +157,7 @@ for example:
 
 ...
 ```
+
+Modify these variables in order to change the colors of the app. In order to modify the appearance in a more complex
+way, you can use the full extent of CSS capabilities - there are no limits here but your imagination. For example,
+`high_contrast_dark.css` extends the CSS in order to add outlines to elements that do not have them by default.
