+++ tests/themes/local_storage_style_saver.test.ts
@@ -1,6 +1,6 @@
 
 import { assert } from "chai";
-import { LocalStorageThemeSaver, ThemeSaver } from "../../source/misc/theme_saver";
+import { LocalStorageThemeSaver, ThemeSaver } from "../../source/themes/theme_saver";
 
 describe (
 	"LocalStorageStyleSaver",
